requirements.txt
djangorestframework
django-cors-headers
django-resized
django-nested-inline

pillow
djoser
pytils
django-cleanup
django-filter
